pyproject.toml
[build-system]
    requires = ["hatchling"]
    build-backend = "hatchling.build"

[project]
    name = "versecbot_smile_back"
    version = "0.2"
    authors = [{ name = "Museus", email = "versecbot@museus.dev" }]
    description = "VerSecBot plugin to return smiles"
    readme = "README.md"
    requires-python = ">=3.8"
    classifiers = [
        "Programming Language :: Python :: 3",
        "Operating System :: OS Independent",
    ]

    dependencies = [
        "discord-py>=2,<3",
        "pydantic_settings>=2,<3",
        "versecbot_interface>=0.1",
    ]

    license = "MIT"
    license-files = ["LICEN[CS]E*"]
    [project.urls]
        Homepage = "https://github.com/Museus/versecbot-SmileBack"
        Issues = "https://github.com/Museus/versecbot-SmileBack/issues"


    [project.optional-dependencies]
        qa = ["ruff", "pytest"]

    [project.entry-points.'versecbot.plugins']
        smile_back = 'versecbot_smile_back:SmileBackPlugin'
